refactor(session): derive the listing tier from the summary tier at subscribe; delete the history import
The memory log imported every historical stream at the root's first
open: one log row per fact, walked inside the synchronous graph build,
and the reason the open had a yield budget to stay under (#11909). The
listing tier is now the summary tier's, read when a reader subscribes,
the way the transcript tier already reads the store for a subscribed
stream: no history pass at open, no history rows held.

Historical means existed when the graph opened. The membership is the
summary tier's key set taken once at the log's build; a stream born
after that enters through its own run.start row alone, so the listing
tier can never mint a live stream at commit 0 ahead of its real row.

Closes #11907
